Repackaged CraftCommons. com.craftfire.commons - classes used only from outside com.craftfire.commons.util - classes used internally as well, that don't deserve an own package
