+++ src/main/java/com/craftfire/commons/FileDownloader.java
@@ -17,7 +17,7 @@
  * You should have received a copy of the GNU Lesser General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package com.craftfire.commons.classes;
+package com.craftfire.commons;
 
 import java.io.BufferedInputStream;
 import java.io.File;
