+++ src/main/java/com/craftfire/commons/PhpSerializer.java
@@ -23,6 +23,8 @@ import java.util.Iterator;
 import java.util.List;
 import java.util.Map;
 
+import com.craftfire.commons.util.Util;
+
 /**
  * PhpSerializer - util class to serialize java objects with php serialization
  *
