+++ src/main/java/com/craftfire/commons/TranslationManager.java
@@ -17,9 +17,9 @@
  * You should have received a copy of the GNU Lesser General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package com.craftfire.commons.managers;
+package com.craftfire.commons;
 
-import com.craftfire.commons.CraftCommons;
+import com.craftfire.commons.util.LoggingManager;
 
 import java.io.DataOutputStream;
 import java.io.IOException;
