+++ src/main/java/com/craftfire/commons/analytics/AnalyticsException.java
@@ -17,11 +17,9 @@
  * You should have received a copy of the GNU Lesser General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package com.craftfire.commons.exceptions;
+package com.craftfire.commons.analytics;
 
-import com.craftfire.commons.Util;
-import com.craftfire.commons.classes.AnalyticsData;
-import com.craftfire.commons.managers.AnalyticsManager;
+import com.craftfire.commons.util.Util;
 
 public class AnalyticsException extends Exception {
     private AnalyticsManager manager;
