+++ src/main/java/com/craftfire/commons/analytics/AnalyticsManager.java
@@ -17,12 +17,10 @@
  * You should have received a copy of the GNU Lesser General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package com.craftfire.commons.managers;
+package com.craftfire.commons.analytics;
 
-import com.craftfire.commons.Util;
-import com.craftfire.commons.classes.AnalyticsData;
-import com.craftfire.commons.classes.AnalyticsItem;
-import com.craftfire.commons.exceptions.AnalyticsException;
+import com.craftfire.commons.util.LoggingManager;
+import com.craftfire.commons.util.Util;
 
 import java.io.*;
 import java.net.*;
