+++ src/main/java/com/craftfire/commons/cache/MetadatableCacheItem.java
@@ -17,8 +17,28 @@
  * You should have received a copy of the GNU Lesser General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package com.craftfire.commons.enums;
+package com.craftfire.commons.cache;
 
-public enum DataType {
-    MYSQL, H2
+import java.util.HashMap;
+import java.util.Map;
+
+public class MetadatableCacheItem extends CacheItem {
+    private Map<String, Object> metaMap = new HashMap<String, Object>();
+
+    public MetadatableCacheItem(Object id, int cacheTime, Object object) {
+        super(id, cacheTime, object);
+    }
+
+    public Object getMetaData(String metaKey) {
+        return this.metaMap.get(metaKey);
+    }
+
+    public void setMetaData(String metaKey, Object metaValue) {
+        this.metaMap.put(metaKey, metaValue);
+    }
+
+    public void removeMetaData(String metaKey) {
+        this.metaMap.remove(metaKey);
+    }
+
 }
