+++ src/main/java/com/craftfire/commons/database/DataType.java
@@ -17,28 +17,8 @@
  * You should have received a copy of the GNU Lesser General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package com.craftfire.commons.classes;
+package com.craftfire.commons.database;
 
-import java.util.HashMap;
-import java.util.Map;
-
-public class MetadatableCacheItem extends CacheItem {
-    private Map<String, Object> metaMap = new HashMap<String, Object>();
-
-    public MetadatableCacheItem(Object id, int cacheTime, Object object) {
-        super(id, cacheTime, object);
-    }
-
-    public Object getMetaData(String metaKey) {
-        return this.metaMap.get(metaKey);
-    }
-
-    public void setMetaData(String metaKey, Object metaValue) {
-        this.metaMap.put(metaKey, metaValue);
-    }
-
-    public void removeMetaData(String metaKey) {
-        this.metaMap.remove(metaKey);
-    }
-
+public enum DataType {
+    MYSQL, H2
 }
