+++ src/main/java/com/craftfire/commons/database/FieldType.java
@@ -17,7 +17,7 @@
  * You should have received a copy of the GNU Lesser General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package com.craftfire.commons.enums;
+package com.craftfire.commons.database;
 
 public enum FieldType {
     STRING, INTEGER, DATE, BLOB, BINARY, BOOLEAN, REAL, NULL, UNKNOWN
