+++ src/main/java/com/craftfire/commons/encryption/Encryption.java
@@ -17,7 +17,7 @@
  * You should have received a copy of the GNU Lesser General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package com.craftfire.commons.enums;
+package com.craftfire.commons.encryption;
 
 public enum Encryption {
     CRC32, MD5, SHA1, SHA256, SHA512, WHIRLPOOL, PHPASS, BLOWFISH
