+++ src/main/java/com/craftfire/commons/ip/IPv6IntervalRange.java
@@ -19,8 +19,8 @@
  */
 package com.craftfire.commons.ip;
 
-import com.craftfire.commons.classes.Version;
-import com.craftfire.commons.classes.VersionRange;
+import com.craftfire.commons.util.Version;
+import com.craftfire.commons.util.VersionRange;
 
 /**
  * An IPRange implementation for IPv6 intervals.
