+++ src/main/java/com/craftfire/commons/util/Util.java
@@ -17,7 +17,7 @@
  * You should have received a copy of the GNU Lesser General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package com.craftfire.commons;
+package com.craftfire.commons.util;
 
 import java.io.IOException;
 import java.net.HttpURLConnection;
